[fix] EditUserInfo 뷰 분리, 색상 적용
