+++ ReptileHub/SceneDelegate.swift
@@ -19,7 +19,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         guard let windowScene = (scene as? UIWindowScene) else { return }
         self.window = UIWindow(windowScene: windowScene)
         
-        self.window?.rootViewController = LoginViewController()
+        self.window?.rootViewController = TabbarViewController()
         self.window?.makeKeyAndVisible()
     }
